removed api.js for simplicity and moved classes to server.js
