+++ index.html
@@ -8,6 +8,6 @@
 </head>
 <body>
     <h1>Rap Names API</h1>
-    <p>Please use https://rapper-api-100devs-class39.herokuapp.com/api/rapper-name</p>
+    <p>Please use https://rapper-api-100devs-class39.herokuapp.com/api/pj</p>
 </body>
 </html>
